services/auth/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
pymongo==4.6.0
motor==3.3.1
python-dotenv==1.0.0
pydantic==2.5.1
email-validator==2.2.0
